+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format of this changelog is based on
 [Keep a Changelog](https://keepachangelog.com/), and this project adheres to
 [Semantic Versioning](https://semver.org/).
 
+## Upcoming
+
+### Fixed
+
+  - `launch!` without rounding now has the correct gap behind the pad
+
 ## 1.0.0 (2025-02-27)
 
 Initial release.